+++ .git-blame-ignore-revs
@@ -12,5 +12,5 @@ b9084ce695c49b5645b82436b19ec6cc3bd30280
 a5d7385a79ef242e7eb1b1bf7b3746b6ebd6a0d0
 3e8cf70c2d3b36d19fcb764599f6dfa6eb924b77
 27f25d037854caa7e15496e4c6fa9d520f365ac9
-# Restyling of evolven3fit_new
+# Restyling of evolven3fit
 07d8d315aff44f98cb546e09088aa981e9778466
