+++ .github/workflows/fitbot.yml
@@ -86,7 +86,7 @@ jobs:
       run: |
         conda activate nnpdfenv
         cd $RUNFOLDER
-        evolven3fit_new evolve $RUNCARD
+        evolven3fit evolve $RUNCARD
     # running postfit
     - name: Postfit and upload fit
       shell: bash -l {0}
