+++ n3fit/src/evolven3fit/cli.py
@@ -1,7 +1,7 @@
 from . import evolve, utils
 
 
-def cli_evolven3fit_new(
+def cli_evolven3fit(
     configuration_folder, q_fin, q_points, op_card_info, theory_card_info, dump, load, force
 ):
     """Evolves the fitted PDFs.
