+++ n3fit/src/n3fit/scripts/evolven3fit.py
@@ -1,12 +1,12 @@
 """
-This module contains the CLI for evolven3fit_new
+This module contains the CLI for evolven3fit
 """
 from argparse import ArgumentParser
 import logging
 import pathlib
 import sys
 
-from evolven3fit_new import cli, eko_utils, evolve
+from evolven3fit import cli, eko_utils, evolve
 import numpy as np
 
 from eko.runner.managed import solve
@@ -89,8 +89,13 @@ def construct_evolven3fit_parser(subparsers):
     return parser
 
 
+def evolven3fit_new():
+    _logger.warning("`evolven3fit_new` is deprecated. Please use `evolven3fit` instead.")
+    main()
+
+
 def main():
-    parser = ArgumentParser(description="evolven3fit_new - a script with tools to evolve PDF fits")
+    parser = ArgumentParser(description="evolven3fit - a script with tools to evolve PDF fits")
     parser.add_argument(
         "-q", "--q-fin", type=float, default=None, help="Final q-value of the evolution"
     )
@@ -125,7 +130,7 @@ def main():
         theory_card_info["use_fhmruvv"] = args.use_fhmruvv
 
     if args.actions == "evolve":
-        cli.cli_evolven3fit_new(
+        cli.cli_evolven3fit(
             args.configuration_folder,
             args.q_fin,
             args.q_points,
